jobmanager: remove job manager v1 code
